requirements.txt
# requirements.txt

python-dotenv
groq
pandas
numpy
python-dotenv
pyyaml
flask
markdown